refactor: deduplicate NDJSON parsing and add delegation macro for agent runners
- Extract shared `parse_ndjson()` into `agents/mod.rs`; remove identical
  copies from `codex.rs` (method) and `opencode.rs` (free function +
  method wrapper). All call sites updated to use `super::parse_ndjson()`.
- Update `router/llm.rs` to call `agents::parse_ndjson()` directly instead
  of the now-removed `opencode::parse_ndjson_events()`.
- Add `delegate_agent_runner!(Type, field)` macro in `mod.rs` that generates
  a fully-forwarding `AgentRunner` impl; replace the 37-line boilerplate
  impl in `kimi.rs` with a single macro invocation.

All 2380 tests pass.
