+++ src/engine/runner/agents/kimi.rs
@@ -9,7 +9,7 @@
 //! output format differences from native Claude.
 
 use super::minimax::MiniMaxClaudeRunner;
-use super::{AgentError, AgentRunner, ParsedResponse, PermissionRules};
+use super::{delegate_agent_runner, AgentError, AgentRunner, ParsedResponse, PermissionRules};
 
 /// Runner for Kimi-via-Claude agents (same output handling as MiniMaxClaude).
 /// Named `KimiClaudeRunner` to distinguish from a future native Kimi CLI runner.
@@ -25,44 +25,7 @@ impl KimiClaudeRunner {
     }
 }
 
-impl AgentRunner for KimiClaudeRunner {
-    #[cfg(test)]
-    fn name(&self) -> &str {
-        self.inner.name()
-    }
-
-    fn build_command(
-        &self,
-        model: Option<&str>,
-        timeout_cmd: &str,
-        sys_file: &str,
-        msg_file: &str,
-        permissions: &PermissionRules,
-    ) -> String {
-        self.inner
-            .build_command(model, timeout_cmd, sys_file, msg_file, permissions)
-    }
-
-    fn parse_response(&self, raw: &str) -> Result<ParsedResponse, AgentError> {
-        self.inner.parse_response(raw)
-    }
-
-    fn extract_text(&self, raw: &str) -> Result<String, AgentError> {
-        self.inner.extract_text(raw)
-    }
-
-    fn classify_error(&self, exit_code: i32, stdout: &str, stderr: &str) -> AgentError {
-        self.inner.classify_error(exit_code, stdout, stderr)
-    }
-
-    fn router_command(
-        &self,
-        prompt: &str,
-        model: Option<&str>,
-    ) -> anyhow::Result<tokio::process::Command> {
-        self.inner.router_command(prompt, model)
-    }
-}
+delegate_agent_runner!(KimiClaudeRunner, inner);
 
 #[cfg(test)]
 mod tests {
